requirements.txt
langchain
langchain-openai
langchain-community
fastapi
uvicorn
streamlit
requests
python-dotenv
yfinance
pydantic
duckduckgo-search